requirement.txt
Flask
psycopg2-binary
python-dotenv
Flask-Cors
gunicorn
#pip install -r requirement.txt